clients/rpi_eink/requirements.txt
pillow
numpy
gpiozero
spidev
dotenv